+++ src/main/java/com/example/moviereserve/dto/venues/VenuesSetupRequestDto.java
@@ -1,4 +1,4 @@
-package com.example.moviereserve.dto;
+package com.example.moviereserve.dto.venues;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
